integrate latest aiopg changes
