+++ tests/contrib/aiopg/test_aiopg.py
@@ -1,8 +1,8 @@
 # stdlib
-import asynctest
 import asyncio
-import time
 import sys
+from textwrap import dedent
+import time
 
 # 3p
 import aiopg
@@ -16,21 +16,24 @@ from ddtrace import Pin
 # testing
 from tests.contrib.config import POSTGRES_CONFIG
 from tests.test_tracer import get_dummy_tracer
+from tests.contrib.asyncio.utils import AsyncioTestCase, mark_asyncio
 
 
 TEST_PORT = str(POSTGRES_CONFIG['port'])
-PY_35 = sys.version_info >= (3, 5)
+PY35 = sys.version_info >= (3, 5)
 
 
-class TestPsycopgPatch(asynctest.TestCase):
+class TestPsycopgPatch(AsyncioTestCase):
     # default service
     TEST_SERVICE = 'postgres'
 
     def setUp(self):
+        super().setUp()
         self._conn = None
         patch()
 
     def tearDown(self):
+        super().tearDown()
         if self._conn and not self._conn.closed:
             self._conn.close()
 
@@ -39,10 +42,9 @@ class TestPsycopgPatch(asynctest.TestCase):
     @asyncio.coroutine
     def _get_conn_and_tracer(self):
         conn = self._conn = yield from aiopg.connect(**POSTGRES_CONFIG)
-        tracer = get_dummy_tracer()
-        Pin.get_from(conn).clone(tracer=tracer).onto(conn)
+        Pin.get_from(conn).clone(tracer=self.tracer).onto(conn)
 
-        return conn, tracer
+        return conn, self.tracer
 
     @asyncio.coroutine
     def assert_conn_is_traced(self, tracer, db, service):
@@ -99,28 +101,36 @@ class TestPsycopgPatch(asynctest.TestCase):
         eq_(span.meta["out.port"], TEST_PORT)
         eq_(span.span_type, "sql")
 
-    @asyncio.coroutine
+    if PY35:
+        # We have to exec this due to syntax errors on earlier versions.
+        exec(dedent("""
+        async def _test_cursor_ctx_manager(self):
+            conn, tracer = await self._get_conn_and_tracer()
+            cur = await conn.cursor()
+            t = type(cur)
+    
+            async with conn.cursor() as cur:
+                assert t == type(cur), "%s != %s" % (t, type(cur))
+                await cur.execute(query="select 'blah'")
+                rows = await cur.fetchall()
+                assert len(rows) == 1
+                assert rows[0][0] == 'blah'
+    
+            spans = tracer.writer.pop()
+            assert len(spans) == 1
+            span = spans[0]
+            eq_(span.name, "postgres.query")
+        """))
+
+    @mark_asyncio
     def test_cursor_ctx_manager(self):
         # ensure cursors work with context managers
         # https://github.com/DataDog/dd-trace-py/issues/228
 
-        conn, tracer = yield from self._get_conn_and_tracer()
-        cur = yield from conn.cursor()
-        t = type(cur)
+        if PY35:
+            yield from self._test_cursor_ctx_manager()
 
-        with (yield from conn.cursor()) as cur:
-            assert t == type(cur), "%s != %s" % (t, type(cur))
-            yield from cur.execute(query="select 'blah'")
-            rows = yield from cur.fetchall()
-            assert len(rows) == 1
-            assert rows[0][0] == 'blah'
-
-        spans = tracer.writer.pop()
-        assert len(spans) == 1
-        span = spans[0]
-        eq_(span.name, "postgres.query")
-
-    @asyncio.coroutine
+    @mark_asyncio
     def test_disabled_execute(self):
         conn, tracer = yield from self._get_conn_and_tracer()
         tracer.enabled = False
@@ -129,7 +139,7 @@ class TestPsycopgPatch(asynctest.TestCase):
         yield from (yield from conn.cursor()).execute("select 'blah'")
         assert not tracer.writer.pop()
 
-    @asyncio.coroutine
+    @mark_asyncio
     def test_manual_wrap_extension_types(self):
         conn, _ = yield from self._get_conn_and_tracer()
         # NOTE: this will crash if it doesn't work.
@@ -137,7 +147,7 @@ class TestPsycopgPatch(asynctest.TestCase):
         #   TypeError: argument 2 must be a connection, cursor or None
         extras.register_uuid(conn_or_curs=conn)
 
-    @asyncio.coroutine
+    @mark_asyncio
     def test_connect_factory(self):
         tracer = get_dummy_tracer()
 
@@ -156,7 +166,7 @@ class TestPsycopgPatch(asynctest.TestCase):
         }
         eq_(service_meta, expected)
 
-    @asyncio.coroutine
+    @mark_asyncio
     def test_patch_unpatch(self):
         tracer = get_dummy_tracer()
         writer = tracer.writer
@@ -197,7 +207,3 @@ class TestPsycopgPatch(asynctest.TestCase):
         spans = writer.pop()
         assert spans, spans
         eq_(len(spans), 1)
-
-
-if __name__ == '__main__':
-    asynctest.main()
