+++ tox.ini
@@ -49,7 +49,7 @@ envlist =
     {py27,py34,py35,py36}-requests{208,209,210,211,212,213}
     {py27,py34,py35,py36}-sqlalchemy{10,11}-psycopg2{27}-mysqlconnector{21}
     {py27,py34,py35,py36}-psycopg2{25,26,27}
-    {py34,py35,py36}-aiopg
+    {py34,py35,py36}-aiopg{012,013}
     {py34,py35,py36}-aiobotocore
     {py27,py34,py35,py36}-redis{26,27,28,29,210}
     {py27,py34,py35,py36}-sqlite3
@@ -96,7 +96,8 @@ deps =
     aiobotocore: asynctest
     aiobotocore: moto>=1.0.1
     aiobotocore: flask
-    aiopg: aiopg
+    aiopg012: aiopg>=0.12,<0.13
+    aiopg013: aiopg>=0.13,<0.14
     aiopg: sqlalchemy
     aiopg: asynctest
     aiohttp12: aiohttp>=1.2,<1.3
